feat: implement netbox auto-updater and notification wrapper

+++ netbox/netbox_autoUpdater.sh
@@ -6,11 +6,10 @@ set -euo pipefail
 NETBOX_DIR="/opt/netbox"
 GITHUB_API="https://api.github.com/repos/netbox-community/netbox/releases/latest"
 
-# Verifica se o link simbólico existe
-if [ ! -L "$NETBOX_DIR" ]; then
-  echo "Erro: $NETBOX_DIR não é um link simbólico."
-  exit 1
-fi
+
+LOG_FOLDER="/var/log/NetboxUpdateLog"
+LOG_FILE="$LOG_FOLDER/Netbox_update.log"
+mkdir -p "$LOG_FOLDER"
 
 # Extrai versão do link simbólico
 REAL_PATH=$(readlink -f "$NETBOX_DIR")
@@ -19,40 +18,50 @@ CURRENT_VERSION=$(basename "$REAL_PATH" | sed 's/^netbox-//')
 # Obtém a versão mais recente do GitHub
 LATEST_VERSION=$(curl -s "$GITHUB_API" | grep '"tag_name":' | head -n 1 | cut -d '"' -f4 | sed 's/^v//')
 
-echo "Versão instalada: $CURRENT_VERSION"
-echo "Última versão disponível: $LATEST_VERSION"
+
+# --- Functions ---
+# Logging function
+log_message() {
+    echo "$(date '+%Y-%m-%d %H:%M:%S'): $1" | tee -a "$LOG_FILE"
+}
+
+# Verifica se o link simbólico existe
+if [ ! -L "$NETBOX_DIR" ]; then
+  log_message "Erro: $NETBOX_DIR não é um link simbólico."
+  exit 1
+fi
+
+log_message "Versão instalada: $CURRENT_VERSION"
+log_message "Última versão disponível: $LATEST_VERSION"
 
 if [ "$CURRENT_VERSION" != "$LATEST_VERSION" ]; then
-    echo "🔔 Uma nova versão do NetBox está disponível!" 
-    NETBOX_UPDATERLOG="/var/log/NetboxUpdateLog"
-    TODAY="$(date '+%Y-%m-%d_%H-%M-%S')"
-    mkdir -p "$NETBOX_UPDATERLOG"
-    echo "🔔 Uma nova versão do NetBox está disponível!" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
-    echo "Iniciando o UPGRADE ⬆️" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "🔔 Uma nova versão do NetBox está disponível!" 
+    
+    log_message "🔔 Uma nova versão do NetBox está disponível!"
+    log_message "Iniciando o UPGRADE ⬆️"
     FROM="$CURRENT_VERSION"
     TO="$LATEST_VERSION"
     NETBOX_DIR="/opt/netbox"
     BACKUP_DIR="/opt/netbox-backups/$(date +%F_%T)$LATEST_VERSION"
     mkdir -p "$BACKUP_DIR"
     
-    echo "Iniciando upgrade de NetBox $FROM para $TO..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
-    echo "⭐ Fazendo backup em $BACKUP_DIR..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "Iniciando upgrade de NetBox $FROM para $TO..."
+    log_message "⭐ Fazendo backup em $BACKUP_DIR..."
     pg_dump -U netbox -h localhost -Fc netbox >> "$BACKUP_DIR/netbox_$FROM.dump"
 
     mkdir -p "$NETBOX_DIR/netbox/media"
     cp -pr "$NETBOX_DIR/netbox/media" "$BACKUP_DIR/"
     cp -r "$NETBOX_DIR/netbox/scripts" "$BACKUP_DIR/"
     cp -r "$NETBOX_DIR/netbox/reports" "$BACKUP_DIR/"
-    echo "✅ Backup concluído com sucesso!" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
-    
+    log_message "✅ Backup concluído com sucesso!"
 
-    echo "⭐ Baixando NetBox v$TO..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "⭐ Baixando NetBox v$TO..."
     wget -q "https://github.com/netbox-community/netbox/archive/v$TO.tar.gz" -O "/tmp/netbox-$TO.tar.gz"
     sudo tar -xzf "/tmp/netbox-$TO.tar.gz" -C /opt
     sudo ln -sfn "/opt/netbox-$TO" "$NETBOX_DIR"
-    echo "✅ Baixado com sucesso!" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "✅ Baixado com sucesso!"
 
-    echo "⭐ Migrando configurações e dados customizados..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "⭐ Migrando configurações e dados customizados..."
     sudo cp "/opt/netbox-$FROM/local_requirements.txt" "$NETBOX_DIR/"
     sudo cp "/opt/netbox-$FROM/netbox/netbox/configuration.py" "$NETBOX_DIR/netbox/netbox/"
     sudo cp "/opt/netbox-$FROM/netbox/netbox/ldap_config.py" "$NETBOX_DIR/netbox/netbox/"
@@ -63,20 +72,18 @@ if [ "$CURRENT_VERSION" != "$LATEST_VERSION" ]; then
     sudo cp -pr "/opt/netbox-$FROM/netbox/media" "$NETBOX_DIR/netbox/"
     sudo rm -rf "$NETBOX_DIR/netbox/static/netbox_topology_views/" && cp -r "/opt/netbox-$FROM/netbox/static/netbox_topology_views" "$NETBOX_DIR/netbox/static/" 
     sudo chown netbox:netbox -R /opt/netbox
-    echo "✅ Migrado com sucesso!" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "✅ Migrado com sucesso!"
 
-    echo "⭐ Executando upgrade.sh..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "⭐ Executando upgrade.sh..."
     cd "$NETBOX_DIR"
     sudo ./upgrade.sh
-    echo "Reiniciando serviços NetBox..." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "Reiniciando serviços NetBox..."
     sudo systemctl restart netbox netbox-rq
     sudo chown netbox:netbox -R /opt/netbox/
-    echo "✅ Upgrade concluído com sucesso!" >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "✅ Upgrade concluído com sucesso!"
 
 else
-    NETBOX_UPDATERLOG="/var/log/NetboxUpdateLog"
-    TODAY="$(date '+%Y-%m-%d_%H-%M-%S')"
-    echo "✅ NetBox está atualizado." >> "$NETBOX_UPDATERLOG/Netbox_update.log"
+    log_message "✅ NetBox está atualizado."
     exit 0
 fi
